+++ classes/evRecur.class.php
@@ -11,7 +11,7 @@
 *   @author     Lee Garner <lee@leegarner.com>
 *   @copyright  Copyright (c) 2011-2016 Lee Garner <lee@leegarner.com>
 *   @package    evlist
-*   @version    1.3.2
+*   @version    1.4.3
 *   @license    http://opensource.org/licenses/gpl-2.0.php
 *               GNU Public License v2 or later
 *   @filesource
@@ -70,7 +70,6 @@ class evRecurBase
         } else {
             $this->duration = 0;      // single day event
         }
-
     }
 
 
@@ -179,8 +178,12 @@ class evRecurBase
     */
     public function storeEvent($start)
     {
+        global $_CONF;
+
         if ($this->duration > 0) {
-            $enddate = strtotime("+$duration day" , strtotime($start)) ;
+            $d = new Date($start, $_CONF['timezone']);
+            $e = $d->Add(new DateInterval("P{$this->duration}D"));
+            $enddate = $e->format('Y-m-d', true);
         } else {
             $enddate = $start;
         }
@@ -195,9 +198,7 @@ class evRecurBase
                         'tm_start2'  => $this->event->time_start2,
                         'tm_end2'    => $this->event->time_end2,
         );
-
     }   // function storeEvent()
-
 
 }   // class evRecurBase
 
@@ -363,7 +364,7 @@ class evRecurMonthly extends evRecurBase
         global $_EV_CONF;
 
         $days_on = $this->event->rec_data['listdays'];
-        if (!is_array($days_on)) return $this->events;
+        if (!is_array($days_on)) return false;
 
         $occurrence = $this->dt_start;
 
@@ -401,14 +402,12 @@ class evRecurMonthly extends evRecurBase
                 if ($this->skip > 0) {
                     $occurrence = $this->SkipWeekend($occurrence);
                 }
-                if ($occurrence != NULL) {
+                if ($occurrence !== NULL) {
                     $this->storeEvent($occurrence);
                     $count++;
                 }
 
                 if ($count > $_EV_CONF['max_repeats']) break;
-
-                //list($y, $m, $d) = explode('-', $occurrence);
 
             }   // foreach days_on
 
@@ -481,7 +480,9 @@ class evRecurYearly extends evRecurBase
 /**
 *   Class to handle weekly recurrences.
 *   This handles multiple occurrences per week, specified by day number.
+*
 *   @package evlist
+*   @return mixed   Array of events on success, False on failure
 */
 class evRecurWeekly extends evRecurBase
 {
@@ -490,6 +491,8 @@ class evRecurWeekly extends evRecurBase
         global $_EV_CONF;
 
         $days_on = $this->event->rec_data['listdays'];
+        if (empty($days_on)) return false;
+
         $occurrence = $this->dt_start;
 
         $num_intervals = count($days_on);