+++ language/french_canadian_utf-8.php
@@ -58,6 +58,7 @@ $LANG_EVLIST = array(
 'err_cal_import'        => 'Il y a eu %d erreurs importation à partir du calendrier. Vérifier votre journal d erreurs pour plus de détails',
 'err_import_event'      => 'Erreur d importation événement %s',
 'err_cal_notavail'      => 'le calendrier plugin les données ne sont pas disponibles. ',
+'err_upd_repeats'   => 'Error updating repeat events',
 'info'              => 'Information',
 'warning'           => 'Avertissement',
 'alerte'            => 'Alerte',
@@ -370,6 +371,8 @@ $LANG_EVLIST = array(
 'timezone'      => 'Timezone',
 'tz_local'      => 'Guest&apos;s local timezone',
 'tz_select'     => 'Select Timezone',
+'msg_item_updated' => 'Item has been updated',
+'msg_item_nochange' => 'Item was not changed',
 );
 
 $PLG_evlist_MESSAGE1         = "Cet événement n\ 't permettre les inscriptions, ou vous n'avez pas accès. ";